Fix build warnings in TradingStrategies
- Fixed time_t format warnings in StrategyUserInterface.c (lines 492, 636)
  Changed %d to %lld with (long long) cast for time_t variables
- Fixed double-to-float conversion warnings in StrategyUserInterface.c (lines 676-680)
  Added explicit (float) casts to atof() return values
- Fixed double-to-int conversion warning in Screening.c (line 560)
  Added explicit (int) cast to trend array assignment

Result: TradingStrategies project now builds with 0 C compiler warnings (only benign MSB8028 build system warnings remain)

+++ core/TradingStrategies/src/StrategyUserInterface.c
@@ -489,7 +489,7 @@ AsirikuyReturnCode saveTradingInfo(int instanceID, Order_Info * pOrderInfo)
 	fprintf(fp, "%f\n", pOrderInfo->openPrice);
 	fprintf(fp, "%f\n", pOrderInfo->stopLossPrice);
 	fprintf(fp, "%f\n", pOrderInfo->takeProfitPrice);
-	fprintf(fp, "%d\n", pOrderInfo->timeStamp);
+	fprintf(fp, "%lld\n", (long long)pOrderInfo->timeStamp);
 	
 
 	fclose(fp);
@@ -633,7 +633,7 @@ AsirikuyReturnCode saveVirutalOrdergInfo(int instanceID, OrderInfo orderInfo)
 	fprintf(fp, "%f\n", orderInfo.openPrice);
 	fprintf(fp, "%f\n", orderInfo.stopLoss);
 	fprintf(fp, "%f\n", orderInfo.takeProfit);
-	fprintf(fp, "%d\n", orderInfo.openTime);
+	fprintf(fp, "%lld\n", (long long)orderInfo.openTime);
 
 
 	fclose(fp);
@@ -673,11 +673,11 @@ int readVirtualOrderInfo(int instanceID, OrderInfo *pOrderInfo)
 	fgets(line, 1024, fp);
 	pOrderInfo->isOpen = atoi(line);
 	fgets(line, 1024, fp);
-	pOrderInfo->openPrice = atof(line);
+	pOrderInfo->openPrice = (float)atof(line);
 	fgets(line, 1024, fp);
-	pOrderInfo->stopLoss = atof(line);
+	pOrderInfo->stopLoss = (float)atof(line);
 	fgets(line, 1024, fp);
-	pOrderInfo->takeProfit = atof(line);
+	pOrderInfo->takeProfit = (float)atof(line);
 	fgets(line, 1024, fp);
 	pOrderInfo->openTime = atoi(line);
